Update stats and add Advent of Code link
This removes some the stats generated by
https://github.com/jstrieb/github-stats, which didn't feel particularly
meaningful or representative. Having looked into the source code, it
looks like it's just based on the file sizes by language in all the
repos I've worked on, rather than the distribution of languages in the
code I've actually written/edited. I believe this is a limitation of the
GitHub API, which is its data source

Replacing it are some WakaTime stats via
https://github.com/anuraghazra/github-readme-stats. That repo also
provides some visualisations of GitHub API stats, but again they don't
feel very useful. The WakaTime stats feel more representative of what I
actually work on the most

This commit also adds a link to my Advent of Code solutions repo, and
formats the code with Prettier, my current default for Markdown